Added Ethnic Group Data

+++ Olympic_graph_generator/ItemModel.cs
@@ -5,10 +5,10 @@ namespace Olympic_graph_generator
     public class ItemModel
     {
         public string Name { get; set; }
-        public int Data { get; set; }
+        public double Data { get; set; }
         public Color Color { get; set; }
 
-        public ItemModel(string name, int data, Color color)
+        public ItemModel(string name, double data, Color color)
         {
             Name = name;
             Data = data;
